requirements.txt
tensorflow-gpu==2.1.0
matplotlib
Markdown==2.6.11
numpy==1.18.3
opencv_python
Pillow
six==1.11.0
joblib==0.14.1
pandas==1.0.3
albumentations==0.4.5
streamlit==0.61.0